backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.9.0

# LiveKit
livekit==0.17.0
livekit-agents==0.8.15
livekit-plugins-deepgram==0.6.5
livekit-plugins-cartesia==0.4.3
livekit-plugins-google==0.3.0

# LangChain + Gemini
langchain==0.3.3
langchain-core==0.3.10
langchain-google-genai==2.0.4
google-generativeai==0.8.3

# Twilio
twilio==9.3.3

# Utilities
httpx==0.27.2
python-multipart==0.0.12
